backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2

# Database
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5

# Data validation
pydantic==2.11.7
pydantic_core==2.33.2

# Environment and configuration
python-dotenv==1.1.1
python-multipart==0.0.20

# Development tools
black==25.1.0
isort==6.0.1

# Async support
anyio==4.10.0